Changed DB of Phonenum from int to Long

+++ PetVaccinationTrackerSystem-Project/Data/Entities/Pet.cs
@@ -31,7 +31,7 @@ namespace PetVaccinationTrackerSystem_Project.Data.Entities
 
         [StringLength(100)]
         public string ColorsAndMarkings { get; set; }
-        public int OwnerPhoneNumber { get; set; }
+        public long OwnerPhoneNumber { get; set; }
 
         [StringLength(300)]
         public string Notes { get; set; }
